Add GsfConstraints.h , collect and explain some the constants used in GSF

+++ Tracking/TrkFitter/TrkGaussianSumFilter/test/testMergeComponents.cxx
@@ -91,13 +91,13 @@ main()
 {
   AlignedDynArray<Component1D, alignment> components(n);
   // Create an array of all components to be merged
-  for (int16_t i = 0; i < n; ++i) {
+  for (int8_t i = 0; i < n; ++i) {
     components[i].mean = input[i].mean;
     components[i].cov = input[i].cov;
     components[i].invCov = input[i].invCov;
     components[i].weight = input[i].weight;
   }
-  std::vector<std::pair<int16_t, int16_t>> mergeOrder =
+  std::vector<std::pair<int8_t, int8_t>> mergeOrder =
     findMerges(components.buffer(), n, 12);
   for (const auto& i : mergeOrder){
     std::cout << "[" << i.first << ", " << i.second << "]" << '\n';
